git-svn-id: svn://127.0.0.1/flaskdump/trunk@3057 386f392f-2c41-1343-9bf7-4b6fedb51b34

+++ mobileapp/FlaskUs/platforms/android/assets/www/js/supplies/supplies.controller.js
@@ -3,10 +3,10 @@
     angular.module('flaskApp')
     .controller('SuppliesCtrl', SuppliesCtrl);
 
-    SuppliesCtrl.$inject = ['$scope', 'SupplyService', '$ionicModal','$location','$flaskUtil','$cookies','$state','$timeout','$ionicListDelegate','$stateParams','$localStorage'];
+    SuppliesCtrl.$inject = ['$scope', 'SupplyService', '$ionicModal','$location','$flaskUtil','$cookies','$state','$timeout','$ionicListDelegate','$stateParams','$localStorage','$ionicHistory','$ionicLoading','$ionicPopup','$ionicScrollDelegate'];
 
     /* @ngInject */
-    function SuppliesCtrl($scope,  SupplyService, $ionicModal,$location,$flaskUtil,$cookies,$state,$timeout,$ionicListDelegate,$stateParams,$localStorage ) {
+    function SuppliesCtrl($scope,  SupplyService, $ionicModal,$location,$flaskUtil,$cookies,$state,$timeout,$ionicListDelegate,$stateParams,$localStorage,$ionicHistory,$ionicLoading,$ionicPopup,$ionicScrollDelegate ) {
 
         $scope.userDataList=[];
         $scope.supplies=[];
@@ -14,19 +14,24 @@
         $scope.MyGameDaysSupply=[];
         $scope.hideItem=false;
         $scope.curreentEventId=$stateParams.currEventId;
+        $scope.goBack = function(){
+            $scope.backToList=true;
+            $ionicHistory.goBack();
+        }
         var userDetail=$cookies.getObject('CurrentUser');
         if(userDetail!=undefined) {
             $scope.userId = userDetail.data.userId;
             $scope.hideCheckBox=false;
-            $scope.agreedToTermsOfUse = userDetail.data.agreedToTermsOfUse;
+            $scope.agreedToTermsOfUse =userDetail.data.isContentAdmin;
         }else{
             $scope.userId=0;
             $scope.hideCheckBox=true
             $scope.agreedToTermsOfUse=false;
+            $scope.backToList=false;
         }
         $scope.currentEventSuppplyKey=$scope.curreentEventId+$scope.userId;
 
-
+        $scope.emptySupplyMessage=false;
 
 
 
@@ -88,15 +93,24 @@
         $scope.toggleItem=function(){
             if($scope.MyGameDayList==true) {
                 $scope.hideItem = !$scope.hideItem;
+
                 $("#FlaskUsListdiv").slideToggle("slow", function () {
                 });
+                $timeout(function () {
+                    $ionicScrollDelegate.resize()
+                },500);
             }
         }
         $scope.toggleMyItem=function(){
             if($scope.MyGameDayList==true){
             $scope.hideMyItem=!$scope.hideMyItem;
+
             $( "#myListdiv" ).slideToggle( "slow", function() {});
-                }
+                $timeout(function () {
+                    $ionicScrollDelegate.resize()
+                },500);
+            }
+
 
         }
         $scope.editList=false;
@@ -118,6 +132,7 @@
             SupplyService.getItemByListId(supplyId).then(function (response) {
                 $scope.MyGameDaysSupply = response;
                 $scope.showSelectedSupplyListItem();
+                $ionicScrollDelegate.scrollTop();
 
                 if ($scope.hideMyItem == false) {
                     $scope.toggleMyItem();
@@ -130,6 +145,9 @@
         $scope.initialize=function() {
             SupplyService.getMySupplyList().then(function (response) {
                 $scope.supplies = response;
+                $scope.emptySupplyMessage=true;
+                showEmptySupplyMessage();
+
                 if($scope.MyGameDayList==true && $scope.selectedGameDaySupplyId!="" &&  $scope.hideCheckBox==false &&  $scope.selectedGameDaySupplyId!=undefined){
                     $scope.showSelectedGameDaySupply($scope.selectedGameDaySupplyId);
                     $scope.getSupplyItemBySupplyId($scope.selectedGameDaySupplyId);
@@ -138,6 +156,19 @@
 
                 }
             });
+        }
+        function showEmptySupplyMessage(){
+            angular.forEach($scope.supplies,function(value,key){
+                var counter=0;
+                if(value.isSystem==false){
+                    counter++;
+                }
+                if(counter==0){
+                    $scope.emptySupplyMessage=true;
+                }else{
+                    $scope.emptySupplyMessage=false;
+                }
+            })
         }
 
         $scope.copyForMyGameDaySupply=function(list){
@@ -170,14 +201,8 @@
              $scope.removeSelectedSupply(list);
          }
         }
-        $scope.addNewSuppliesList=function(){
-            setTimeout(setFocus, 50);
-            function setFocus(){
-                document.getElementById("supplyEdit").focus();
-            }
-        }
+
         $scope.saveList = function(list) {
-            // $scope.listItem.unshift( $scope.createdListItem.data);
 
             if ($scope.islistCreated != true) {
                 SupplyService.addSupplies(list,$scope.addAsAdmin).then(function(response){
@@ -206,26 +231,19 @@
 
         }
         $scope.editSupply=function(data){
+
+            angular.forEach($scope.supplies,function(key,value){
+               key.edit=false;
+            })
             data.edit=true;
             $scope.deleteSuplies=true;
             $scope.editSuply=true;
             $ionicListDelegate.closeOptionButtons();
-            setTimeout(setFocus, 50);
-     function setFocus(){
-         if(data.isSystem==false){
-    document.getElementById("editBox").focus();
-        }else{
-             document.getElementById("systemEditBox").focus();
-         }
-     }
+
     }
         $scope.editSupplyItem=function(data){
             data.editItem=true;
-            setTimeout(setFocus, 50);
-            function setFocus(){
-                document.getElementById("supplyEdit").blur();
-                document.getElementById("ItemEditBox").focus();
-            }
+            $ionicListDelegate.closeOptionButtons();
         }
 
         $scope.cancelAdding=function(){
@@ -233,7 +251,7 @@
             $scope.suppliesName="";
         }
         $scope.saveSupply=function(data){
-            if(data.listName!="") {
+            if(data.supplyListName!="") {
                 $scope.deleteSuplies=false;
                 $scope.editSuply=false;
                 SupplyService.updateSupplyList(data.supplyListId,data.supplyListName,data.isSystem).then(function(response){
@@ -245,9 +263,11 @@
                 });
             }else{
                 if(data.isSystem==false) {
-                    document.getElementById("editBox").focus();
+                    $ionicLoading.show({ template: 'List name should not be empty', noBackdrop: false, duration: 1000 });
+
                 }else{
-                    document.getElementById("systemEditBox").focus();
+                    $ionicLoading.show({ template: 'List name should not be empty', noBackdrop: false, duration: 1000 });
+
                 }
             }
         }
@@ -262,7 +282,7 @@
                         data.editItem = false;
                         data.supplyItemId=response.supplyItemId;
                         if($scope.createdListItem.data[$scope.createdListItem.data.length-1].itemName=="") {
-                            setTimeout(setFocusOnItemBox, 50);
+
                         }
                     }
                     else{
@@ -277,25 +297,41 @@
                     data.editItem = false;
                 })
             }
-            else{
-                document.getElementById("ItemEditBox").focus();
+            else if($scope.backToList==true){
+
+            }
+            else {
+                $ionicLoading.show({ template: 'Item name should not be empty', noBackdrop: false, duration: 1000 });
+
             }
         }
         $scope.createdListItem={"data":
             []
         };
         $scope.deleteCreatedItem=function(index,id){
-            SupplyService.deleteSupplyItemById(id).then(function(response){
-                if(response){
-                    $scope.createdListItem.data.splice(index,1);
+            var confirmPopup = $ionicPopup.confirm({
+                title: 'Delete Item?'
+
+            });
+            confirmPopup.then(function(res) {
+                if(res) {
+                    SupplyService.deleteSupplyItemById(id).then(function(response){
+                        if(response){
+                            $scope.createdListItem.data.splice(index,1);
+                        }
+                    });
+
+                } else {
                 }
             });
+
 
         }
         var listItemEmpty={itemName: "",editItem:true}
         $scope.addItem=function(listName,data){
             if (listName == undefined || listName == "") {
-                setTimeout(setFocus, 50);
+                $ionicLoading.show({ template: 'List name should not be empty', noBackdrop: false, duration: 1000 });
+
             }else {
                 SupplyService.addSupplies(listName, $scope.addAsAdmin).then(function (response) {
                     if (response.supplyListId > 0) {
@@ -311,10 +347,12 @@
             }
         $scope.createItem = function (listName,data) {
                 if (listName == undefined || listName == "") {
-                    setTimeout(setFocus, 50);
+                    $ionicLoading.show({ template: 'Item name should not be empty', noBackdrop: false, duration: 1000 });
+
                 }
                 else if (data != undefined && data.itemName == "") {
-                    setTimeout(setFocusOnItemBox, 50);
+                    $ionicLoading.show({ template: 'Item name should not be empty', noBackdrop: false, duration: 1000 });
+
                 }
                 else {
                     $scope.createdListItem.data.push(angular.copy(listItemEmpty));
@@ -322,9 +360,7 @@
                 }
         }
 
-        function setFocusOnItemBox(){
-            document.getElementById("ItemEditBox").focus();
-        }
+
         function setFocus(){
             document.getElementById("supplyEdit").focus();
         };
@@ -359,6 +395,7 @@
                 if(response1.userId>0) {
 
                     var failed=false;
+
                     angular.forEach($scope.itemValue,function(value,key){
                         $scope.supplyItemNames.push(value.supplyItemName);
 
@@ -371,7 +408,9 @@
                     if(failed==true){
                         $flaskUtil.alert("failed to copy");
                     }else{
-                        $scope.supplies.push(list);
+
+                        $scope.supplies.push(response1);
+                        showEmptySupplyMessage();
                     }
 
                 }else{
@@ -390,14 +429,26 @@
         }
 
         $scope.deleteItem=function(index,supplyId){
-            $scope.deleteSuplies=true;
-            SupplyService.deleteSupplyListById(supplyId).then(function(response){
-                if(response){
-                    $ionicListDelegate.closeOptionButtons();
-                    $scope.supplies.splice(index,1);
+            var confirmPopup = $ionicPopup.confirm({
+                title: 'Delete List?'
+            });
+            confirmPopup.then(function(res) {
+                if(res) {
+                    $scope.deleteSuplies=true;
+                    SupplyService.deleteSupplyListById(supplyId).then(function(response){
+                        if(response){
 
+                            $ionicListDelegate.closeOptionButtons();
+                            $scope.supplies.splice(index,1);
+                            showEmptySupplyMessage();
+
+                        }
+                    })
+
+                } else {
                 }
-            })
+            });
+
 
         }
     }
